Fix namespace specification
In the original implementation, there was a mistake in specifying the
namespace, and if the requirement of xx succeeded, the following error
occurred.

```
NameError: uninitialized constant Thor::DidYouMean::SpellChecker
```

Defining `DidYouMean` namespace in Thor is a source of confusion, so I
fixes with delete it.

In addition, `did_you_mean` is added to Gemfile together. Without this, the
test with `did_you_mean` can not be performed correctly.
(Originally it was operating without this specification, rvm installs gem in global)

Also, depending on the version of `DidYouMean`x to be used,
`DidYouMean::SpellChecker` may not be defined. So, I fixed it so that
processing can continue even if `NameError` occurs.

Fixes #635